+++ app/dashboard/layout.tsx
@@ -2,7 +2,7 @@ import { redirect } from "next/navigation";
 import { getUser } from "@/lib/services/users/getUser";
 import UserProvider from "./UserProvider";
 import { routes } from "@/utils/routes/route";
-import { Aside } from "@/components/Dashboard/Aside/Aside";
+import { Aside } from "@/components/dashboard/aside/Aside";
 import { ButtonSwitchTheme } from "@/components/ui/ButtonSwitchTheme";
 
 export default async function DashboardLayout({