+++ config/fonts.ts
@@ -3,4 +3,5 @@ import { Elms_Sans } from "next/font/google";
 export const fontElms = Elms_Sans({
   subsets: ["latin"],
   variable: "--font-elms_sans",
+  fallback: [],
 });